Add files via upload
One of the first small projects I created.  The game does not have a true AI, so the computer chooses a random empty square at the end of the userInput function.  

The game works by setting each square to a value of 0.  When the user chooses (X), the value of the square changes to 1.  When the computer chooses (O), the value of the square changes to 4.  The winner is determined by comparing the rows and columns of squares against a list of logic operands.  The function winner checks periodically for the win condition, so that when the win condition is met, the value of the squares becomes 10 to end the game.  The function newGame simply resets the conditions to default.

Feel free to leave suggestions for improvement.  I viewed a tic-tac-toe game that utilized multidimensional arrays, but was unfamiliar with how the author used them, so I went with a simpler approach.